requirements.txt
tensorflow-cpu==2.1.0
Keras==2.2.5
numpy==1.18.1
glob2==0.7
h5py==2.8.0
